Major update: cloak slot, consumable buffs, pixel art assets, and bug fixes
- Add 5th equipment slot (cloak) across all models, views, and sync
- Add consumable buff tracking (EXP boost, gold boost, luck elixir, etc.)
- Add armor proficiency system per class line
- Add EquipmentQuirk model and crafting materials catalog
- New pixel-art equipment assets (2x/3x) for all weapon and accessory types
- New dungeon biome images (aquatic, celestial, crystal, desert, grove, etc.)
- Add consumable item assets and materials assets
- Fix meditation freeze: remove mutating checkMeditationStreak() from computed property
- Fix splash screen rushing past: enforce 2.5s minimum display time
- Fix arena hasFreeArenaAttempt causing re-render loop (same pattern as meditation)
- Include cloak slot in hero power, effective stats, and stat breakdowns
- Update gold rewards for meditation and level-up

Made-with: Cursor

+++ CouplesQuest/Services/ContentManager.swift
@@ -28,9 +28,9 @@ struct ContentEquipment: Codable, Identifiable, Hashable {
     let baseType: String        // "sword", "axe", "plate", "ring", "cloak", etc.
     let rarity: String          // "common" ... "legendary"
     let primaryStat: String     // "strength", "wisdom", etc.
-    let statBonus: Int
+    let statBonus: Double
     let secondaryStat: String?
-    let secondaryStatBonus: Int
+    let secondaryStatBonus: Double
     let levelRequirement: Int
     let imageName: String?
     let isSetPiece: Bool
@@ -61,9 +61,9 @@ struct ContentMilestoneGear: Codable, Identifiable, Hashable {
     let baseType: String
     let rarity: String
     let primaryStat: String
-    let statBonus: Int
+    let statBonus: Double
     let secondaryStat: String?
-    let secondaryStatBonus: Int
+    let secondaryStatBonus: Double
     let levelRequirement: Int
     let characterClass: String
     let goldCost: Int
